Change quantity field type from decimal to integer in ComboDish model and migration; add migration to alter existing database structure

+++ database/migrations/2026_03_31_000003_create_combo_dishes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('combo_id')->constrained('combos')->cascadeOnDelete();
             $table->foreignId('dish_id')->constrained('dishes')->cascadeOnDelete();
-            $table->decimal('quantity', 10, 2)->default(1);
+            $table->integer('quantity')->default(1);
             $table->integer('sort_order')->default(0);
             $table->boolean('is_active')->default(true);
             $table->timestamps();
